增加userService

+++ src/main/java/com/example/demo/repository/QuestionRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.domain.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question,Long> {
+public interface QuestionRepository extends JpaRepository<Question,Integer> {
 }
